// src/shemas/intent/intent.schema.ts
import { z } from "zod";

export const intentSchema = z.object({
  id: z.string(),
  object: z.string(),
  label: z.string(),
  amount: z.number(),
  receivedAmount: z.number().nullable(),
  currency: z.string(),
  clientSecret: z.string(),
  paymentMethodTypes: z.any(),
  status: z.string(),
  description: z.string().nullable(),
  receiptEmail: z.string().email().nullable(),
  payoutId: z.string().nullable(),
  payoutStatus: z.string().nullable(),
  chargeId: z.string().nullable(),
  createdAt: z.string(),
  updatedAt: z.string(),
  charges: z.array(
    z.object({
      id: z.string(),
      object: z.string(),
      intentId: z.string(),
      financialTransactionId: z.string().nullable(),
      transactionId: z.string(),
      amount: z.number(),
      captured: z.boolean(),
      refunded: z.boolean(),
      disputed: z.boolean(),
      currency: z.string(),
      status: z.string(),
      failureCode: z.string().nullable(),
      failureMessage: z.string().nullable(),
      externalId: z.string().nullable(),
      paymentMethodId: z.string().nullable(),
      createdAt: z.string(),
      updatedAt: z.string(),
      paymentMethod: z.object({
        id: z.string(),
        type: z.string(),
        msisdn: z.string(),
        country: z.string(),
        operator: z.string(),
        createdAt: z.string(),
        updatedAt: z.string(),
        customer: z.object({
          id: z.string(),
          firstName: z.string().nullable(),
          lastName: z.string().nullable(),
          email: z.string().email().nullable(),
          phone: z.string().nullable(),
          accountId: z.string(),
          paymentMethodId: z.string().nullable(),
          createdAt: z.string(),
          updatedAt: z.string(),
        }),
      }),
    }),
  ),
  walletTransaction: z.object({
    id: z.string(),
    object: z.string(),
    walletId: z.string(),
    chargeId: z.string(),
    intentId: z.string(),
    amount: z.number(),
    amountCaptured: z.number(),
    amountFee: z.number(),
    currency: z.string(),
    oldBalance: z.number(),
    time: z.string(),
    date: z.string(),
    newBalance: z.number(),
    createdAt: z.string(),
    updatedAt: z.string(),
  }),
});